+++ __tests__/node/services/middleware/authenticator.test.ts
@@ -1,5 +1,5 @@
-import { Authenticator } from "@/middlewareSupport/authenticator/implementation";
-import { IAuthenticator } from "@/middlewareSupport/authenticator/interface";
+import { Authenticator } from "@/services/middleware/authenticator/implementation";
+import { IAuthenticator } from "@/services/middleware/authenticator/interface";
 import { NextRequest } from "next/server";
 import { getToken } from "next-auth/jwt";
 jest.mock("next-auth/jwt")
