+++ src/services/middleware/authenticator/interface.ts
