+++ src/services/middleware/handler/factory.ts
@@ -1,5 +1,5 @@
-import { IRouteHandler } from "@/middlewareSupport/handler/interface"
-import { RouteHandler } from "@/middlewareSupport/handler/implementation"
+import { IRouteHandler } from "@/services/middleware/handler/interface"
+import { RouteHandler } from "@/services/middleware/handler/implementation"
 import { NextRequest } from "next/server"
 
 export function handlerFactory(request: NextRequest, protectedEndpoint: string): IRouteHandler {
